Add a theme that automatically switches according to system preferences
Android now offers a system-wide dark theme, as well as the option to automatically toggle it based on sunset/sunrise.

This creates a new "AUTO" theme that chooses between `LIGHT` and `BLACK_BREEZE` based on whether the global system night mode is on.

Also makes this new theme the default, so that users will be exposed to the beautiful light theme we have as well.